+++ src/stdio.ts
@@ -17,7 +17,8 @@ import minimist from 'minimist';
 
 import { log } from './logger.js';
 import { ApifyMcpServer } from './mcp-server.js';
-import { getActorDiscoveryTools, getActorAutoLoadingTools } from './tools.js';
+import { getActorDiscoveryTools } from './toolkits/discovery-tools.js';
+import { getActorAutoLoadingTools } from './toolkits/actor-auto-loading-tools.js';
 
 log.setLevel(log.LEVELS.ERROR);
